list

+++ cli/openapi/src/models/file_download.rs
@@ -18,8 +18,9 @@ pub struct FileDownload {
     pub id: Option<serde_json::Value>,
     #[serde(rename = "mnemonic")]
     pub mnemonic: String,
+    /// The type of the inode
     #[serde(rename = "type")]
-    pub r#type: f64,
+    pub r#type: i32,
     #[serde(rename = "name")]
     pub name: String,
     #[serde(rename = "tag", deserialize_with = "Option::deserialize")]
@@ -37,7 +38,7 @@ pub struct FileDownload {
 }
 
 impl FileDownload {
-    pub fn new(id: Option<serde_json::Value>, mnemonic: String, r#type: f64, name: String, tag: Option<String>, data: models::ChunkData, parent_id: Option<serde_json::Value>, created_at: String, updated_at: String, keys: Vec<models::Key>) -> FileDownload {
+    pub fn new(id: Option<serde_json::Value>, mnemonic: String, r#type: i32, name: String, tag: Option<String>, data: models::ChunkData, parent_id: Option<serde_json::Value>, created_at: String, updated_at: String, keys: Vec<models::Key>) -> FileDownload {
         FileDownload {
             id,
             mnemonic,
