+++ cli/openapi/src/models/inode_tree.rs
@@ -18,8 +18,9 @@ pub struct InodeTree {
     pub id: Option<serde_json::Value>,
     #[serde(rename = "mnemonic")]
     pub mnemonic: String,
+    /// The type of the inode
     #[serde(rename = "type")]
-    pub r#type: f64,
+    pub r#type: i32,
     #[serde(rename = "name")]
     pub name: String,
     #[serde(rename = "tag", deserialize_with = "Option::deserialize")]
@@ -41,7 +42,7 @@ pub struct InodeTree {
 }
 
 impl InodeTree {
-    pub fn new(id: Option<serde_json::Value>, mnemonic: String, r#type: f64, name: String, tag: Option<String>, parent_id: Option<serde_json::Value>, created_at: String, updated_at: String, members: Vec<models::Member>, keys: Vec<models::Key>) -> InodeTree {
+    pub fn new(id: Option<serde_json::Value>, mnemonic: String, r#type: i32, name: String, tag: Option<String>, parent_id: Option<serde_json::Value>, created_at: String, updated_at: String, members: Vec<models::Member>, keys: Vec<models::Key>) -> InodeTree {
         InodeTree {
             id,
             mnemonic,
